rewrote track index reading/writing
Added db/file/Index.hpp and db/file/Generic.hpp which implement reading and writing the database's current index object. The Generic class provides facilities useful to other objects that follow in the future. Renamed adb-add to db-add and added proper timestamping as well as a track listing function.

+++ source/tools/db-add.cpp
@@ -2,7 +2,7 @@
 #include <common/sha1.hpp>
 #include <common/uuid.hpp>
 #include <common/Date.hpp>
-#include <db/Tracks.hpp>
+#include <db/file/Index.hpp>
 #include <iostream>
 #include <boost/program_options.hpp>
 #include <boost/filesystem.hpp>
@@ -10,6 +10,7 @@
 #include <vector>
 #include <sstream>
 #include <ctime>
+#include <set>
 
 namespace po = boost::program_options;
 namespace fs = boost::filesystem;
@@ -19,6 +20,7 @@ using std::cerr;
 using std::endl;
 using std::string;
 using std::vector;
+using std::set;
 using std::stringstream;
 
 class DatabaseTool
@@ -87,13 +89,17 @@ protected:
 	po::options_description options;
 	po::options_description hidden_options;
 	po::positional_options_description positional_options;
+
+	string nice_hash(const string& in)
+	{
+		return in.substr(0, 7);
+	}
 };
 
 class ThisTool : public DatabaseTool
 {
 public:
 	vector<string> opt_files;
-	bool opt_import_duplicates;
 
 	ThisTool(int argc, char **argv): DatabaseTool(argc, argv) {}
 	const char* usage_string() { return "file1 ..."; }
@@ -102,7 +108,8 @@ public:
 	{
 		po::options_description parameters("Parameters");
 		parameters.add_options()
-			("duplicates,d", "import duplicate files");
+			("duplicates,d", "import duplicate files")
+			("list,l", "list tracks in database");
 		options.add(parameters);
 
 		hidden_options.add_options()
@@ -121,7 +128,8 @@ public:
 
 	int main()
 	{
-		opt_import_duplicates = vm.count("duplicates");
+		bool opt_import_duplicates = vm.count("duplicates");
+		bool opt_list = vm.count("list");
 		int duplicates_skipped = 0; // increased for every duplicate that was skipped
 
 		// Useful paths
@@ -132,7 +140,7 @@ public:
 
 		// Read the tracks ref if it exists and read the tracks list it points
 		// to. Otherwise create a new empty tracks list.
-		auris::db::Tracks tracks;
+		auris::db::file::Index tracks;
 		string tracks_ref;
 		fs::path tracks_ref_path = dir/"refs"/"tracks";
 		if (fs::exists(tracks_ref_path)) {
@@ -143,7 +151,17 @@ public:
 			tracks.read(ft);
 		}
 
+		// List all tracks and abort if so requested.
+		if (opt_list) {
+			cerr << "# index " << nice_hash(tracks_ref) << " (" << tracks.date << ")\n";
+			for (set<string>::const_iterator it = tracks.tracks.begin(); it != tracks.tracks.end(); it++) {
+				cout << nice_hash(*it) << "\n";
+			}
+			return 0;
+		}
+
 		// Process each input file individually.
+		bool tracks_modified = false;
 		for (int i = 0; i < opt_files.size(); i++)
 		{
 			fs::path path(opt_files[i]);
@@ -173,7 +191,7 @@ public:
 			stringstream meta;
 			meta << "ID: " << uuid << "\n";
 			meta << "Title: " << path.filename().native() << "\n";
-			meta << "Added: " << date << "\n";
+			meta << "Added: " << auris::Date().str() << "\n";
 			meta << "\n" << file_hash << " -- " << path.filename().native() << "\n";
 			string metadata = meta.str();
 
@@ -181,6 +199,7 @@ public:
 			// data to tisk.
 			string meta_hash = auris::sha1().from_string(metadata).hex();
 			tracks.tracks.insert(meta_hash);
+			tracks_modified = true;
 			fs::path db_meta_path = dir/"objects"/meta_hash;
 
 			if (!fs::exists(db_file_path)) fs::copy(path, db_file_path);
@@ -193,28 +212,35 @@ public:
 		}
 
 		// Write the new tracks file.
-		stringstream tracks_stream;
-		tracks.write(tracks_stream);
-		string tracks_string = tracks_stream.str();
-		string tracks_hash = auris::sha1().from_string(tracks_string).hex();
+		if (tracks_modified) {
+			tracks.date = auris::Date().str();
+			tracks.base = tracks_ref;
+			stringstream tracks_stream;
+			tracks.write(tracks_stream);
+			string tracks_string = tracks_stream.str();
+			string tracks_hash = auris::sha1().from_string(tracks_string).hex();
 
-		if (tracks_hash != tracks_ref) {
-			std::ofstream ftracks((dir/"objects"/tracks_hash).c_str());
-			ftracks << tracks_string;
-			ftracks.close();
+			if (tracks_hash != tracks_ref) {
+				std::ofstream ftracks((dir/"objects"/tracks_hash).c_str());
+				ftracks << tracks_string;
+				ftracks.close();
 
-			std::ofstream ftr(tracks_ref_path.c_str());
-			ftr << tracks_hash << "\n";
-			ftr.close();
+				std::ofstream ftr(tracks_ref_path.c_str());
+				ftr << tracks_hash << "\n";
+				ftr.close();
 
-			if (!tracks_ref.empty()) {
-				fs::remove(dir/"objects"/tracks_ref);
+				if (!tracks_ref.empty()) {
+					fs::remove(dir/"objects"/tracks_ref);
+				}
 			}
 		}
 
 		// Print a helpful message in case a duplicate was skipped.
 		if (duplicates_skipped) {
-			cerr << "# Some files were not imported since they already exist in the library or were already added to the index. Add the --duplicates option to force them to be added.\n";
+			cerr <<
+			"# Some files were not imported since they already exist in the library or were\n"
+			"# already added to the index. Add the --duplicates option to force them to be\n"
+			"# added.\n";
 		}
 
 		return 0;
